// frontend/src/client/models/ExtractPageRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Request to extract data from a single URL.
 */
export type ExtractPageRequest = {
    client_id: string;
    url: string;
    crawl_run_id?: (string | null);
};